weather-data-service/requirements.txt
# Core Dependencies for Weather Data Collection Service

# HTTP Client for NOAA API
aiohttp==3.9.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Configuration
python-dotenv==1.0.0

# Type hints (built into Python 3.9+)
typing-extensions==4.8.0

# Optional: For future API development
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Optional: For future database migrations
alembic==1.13.1

# System monitoring and maintenance
psutil==5.9.6
requests==2.31.0

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0